+++ docs/api/changelog.rst
@@ -41,6 +41,9 @@ Fixed
   :class:`imod.mf6.SingleLayerHorizontalFlowBarrierResistance` and other HFB
   packages would have resistances that were double the expected value with
   xugrid >= 0.14.2
+- Fixed bug where :meth:`imod.mf6.SourceSinkMixing.from_flow_model` would return
+  an error upon adding a package which cannot have a ``concentration``, such as 
+  :class:`imod.mf6.HorizontalFlowBarrierResistance`.
 
 Changed
 ~~~~~~~